WebApp/requirements.txt
flask
sqlalchemy
datetime
numpy
pandas
requests
flask-wtf
wtforms
mysql-connector
